refactor: move config/endpoint CLI helpers to config.clihelpers
load_config_or_exit, resolve_endpoints, build_endpoint_filter now live
in config/clihelpers.py. Delete the old clihelpers/config.py and
clihelpers/endpoints.py shims. clihelpers/ is now a pure leaf module
(OutputFormat + StepProgressBar only).
