+++ nbkp/credentials/cli/keyring_status_cmd.py
@@ -11,7 +11,8 @@ from rich.console import Console
 from rich.table import Table
 from rich.text import Text
 
-from ...clihelpers import OutputFormat, load_config_or_exit
+from ...clihelpers import OutputFormat
+from ...config.clihelpers import load_config_or_exit
 from ...config import Config
 from .. import CredentialError, retrieve_passphrase
 from . import app
