+++ nbkp/disks/cli/setup_auth_cmd.py
@@ -7,7 +7,7 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import load_config_or_exit
+from ...config.clihelpers import load_config_or_exit
 from ..auth import POLKIT_RULES_PATH, SUDOERS_RULES_PATH, generate_auth_rules
 from . import app
