+++ nbkp/disks/cli/status_cmd.py
@@ -7,9 +7,9 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import OutputFormat, load_config_or_exit
+from ...clihelpers import OutputFormat
 from ...clihelpers import StepProgressBar
-from ...clihelpers.endpoints import resolve_endpoints
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from . import app
 from .helpers import _probe_volume_status, _show_status_table, _unmanaged_statuses
