+++ nbkp/disks/cli/umount_cmd.py
@@ -7,9 +7,9 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import OutputFormat, load_config_or_exit
+from ...clihelpers import OutputFormat
 from .progress import DisksProgressBar
-from ...clihelpers.endpoints import resolve_endpoints
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from ..lifecycle import UmountResult, mount_volume_count, umount_volumes
 from ..output import volume_display_name
