+++ nbkp/ordering/cli/graph_cmd.py
@@ -9,7 +9,8 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import OutputFormat, load_config_or_exit
+from ...clihelpers import OutputFormat
+from ...config.clihelpers import load_config_or_exit
 from ..output import (
     build_graph_json,
     print_mermaid_ascii_graph,
