+++ nbkp/preflight/cli/troubleshoot_cmd.py
@@ -8,8 +8,7 @@ from typing import Annotated, Optional
 import typer
 
 from ...disks.cli.helpers import managed_mount
-from ...clihelpers import load_config_or_exit
-from ...clihelpers.endpoints import resolve_endpoints
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from ...preflight.output import print_human_troubleshoot
 from . import app
