+++ nbkp/run/cli/run_cmd.py
@@ -11,9 +11,9 @@ from rich.console import Console, Group
 from rich.panel import Panel
 from rich.text import Text
 
-from ...clihelpers import OutputFormat, load_config_or_exit
+from ...clihelpers import OutputFormat
 from ...clihelpers import StepProgressBar
-from ...clihelpers.endpoints import resolve_endpoints
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from ...ordering.output import build_rich_tree_sections
 from ...preflight import PreflightResult
