+++ nbkp/sh/cli/sh_cmd.py
@@ -9,8 +9,7 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import load_config_or_exit
-from ...clihelpers.endpoints import resolve_endpoints
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from .. import ScriptOptions, generate_script
