+++ nbkp/snapshots/cli/show_cmd.py
@@ -8,8 +8,8 @@ from typing import Annotated, Optional
 
 import typer
 
-from ...clihelpers import OutputFormat, load_config_or_exit
-from ...clihelpers.endpoints import resolve_endpoints
+from ...clihelpers import OutputFormat
+from ...config.clihelpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
 from ..cmd_handler import show_all_syncs
 from ..output import print_human_show_results
