+++ tests/preflight/test_cli.py
@@ -27,7 +27,7 @@ from tests.clihelpers import (
 
 class TestLocationValidation:
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_unknown_location_rejected(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -41,7 +41,7 @@ class TestLocationValidation:
         assert "travel" in result.output
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_unknown_exclude_location_rejected(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -62,7 +62,7 @@ class TestLocationValidation:
         assert "--exclude-location" in result.output
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_known_location_accepted(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -92,7 +92,7 @@ class TestLocationValidation:
         )
         assert result.exit_code == 0
 
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_location_on_config_without_locations(self, mock_load: MagicMock) -> None:
         mock_load.return_value = sample_config()
         result = runner.invoke(
@@ -104,7 +104,7 @@ class TestLocationValidation:
 
 class TestCheckCommand:
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_human_output_inactive(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -119,7 +119,7 @@ class TestCheckCommand:
         assert "Preflight" in result.output
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_human_output_all_active(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -133,7 +133,7 @@ class TestCheckCommand:
         assert result.exit_code == 0
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_json_output_inactive(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -153,7 +153,7 @@ class TestCheckCommand:
         assert "syncs" in data
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_json_output_all_active(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -173,7 +173,7 @@ class TestCheckCommand:
         assert "syncs" in data
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_sentinel_only_exit_0_by_default(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
@@ -187,7 +187,7 @@ class TestCheckCommand:
         assert result.exit_code == 0
 
     @patch("nbkp.preflight.cli.helpers.check_all_syncs")
-    @patch("nbkp.clihelpers.config.load_config")
+    @patch("nbkp.config.clihelpers.load_config")
     def test_sentinel_only_exit_1_when_strict(
         self, mock_load: MagicMock, mock_checks: MagicMock
     ) -> None:
